[XLA] Remove a bogus invalid argument message printed out when --v>=1.
When running any trivial XLA program with --v=1, you will see bogus message such
as "Invalid argument: Shape f32[] size may overflow int64". The reason for this
is because in ShapeUtil::ValidateShapeSize, we incorrectly construct an
InvalidArgument object prematurely. This change postpones the construction of
the InvalidArgument object until an invalid argument is actually discovered.

PiperOrigin-RevId: 202959886