Fix crash when running with --v=2.
When doing multi-output fusion and using sibling fusion, it can happen that we
don't need to clone the 'instruction_to_fuse' argument. Right now, we clone,
and then delete the clone again, and at the end of the function try to print
the debug string for the clone (which then crashes).
Instead, we can simply not generate the clone if it is not needed, and catch
this case before printing the debug string.

PiperOrigin-RevId: 203733796